Updates now use existing json as a base
Before, we simply re-posted the file json when updating existing entities. This
sometimes causes issues when, for example, there are system properties that
are set post entity creation.

Example: Creating a Social IdP, which has an ACTIVE status set. On the next run
of the import, when updating the IdP, we inadvertently caused the IdP to enter
into an INACTIVE state.

RESOLVES OKTA-122816
8ac0n: test

OKTA-122095
<<<Jenkins Check-In of Tested SHA: 138ce5dca4f793d16dc0d938d7bc603de93ef6d2 for ci-admin@okta.com>>>
Artifact: salamander-migration